refactor(research): extract reusable content integrity analysis
